+++ packages/ts-logkit/src/formatters/browser.ts
@@ -1,10 +1,5 @@
 import { Event, Formatter, FormattedOutput, Level } from "../types";
 
-/**
- * TODO: Probably delete this? It will be left here for posterity as it is a
- * good example of a formatter for browser environments using %c styling (with css?).
- */
-
 const browserColors: Record<Level, string> = {
   trace: "color: gray",
   debug: "color: cyan",
@@ -14,6 +9,12 @@ const browserColors: Record<Level, string> = {
   fatal: "color: white; background-color: red; font-weight: bold",
 };
 
+/**
+ * NOT SUPPORTED FOR NOW
+ *
+ * TODO: Probably delete this? It will be left here for posterity as it is a
+ * good example of a formatter for browser environments using %c styling (with css?).
+ */
 export function formatBrowser(event: Event): [string, string, ...unknown[]] {
   const style = browserColors[event.level] || "";
   const time = new Date(event.timestamp).toISOString();
@@ -33,8 +34,6 @@ export function formatBrowser(event: Event): [string, string, ...unknown[]] {
  * Browser formatter for browser environments
  * Uses CSS styling via %c format specifier
  */
-export class BrowserFormatter implements Formatter {
-  format(event: Event): FormattedOutput {
-    return formatBrowser(event);
-  }
-}
+export const browserFormatter: Formatter = {
+  format: formatBrowser,
+};