+++ packages/ts-logkit/src/noop.ts
@@ -1,8 +1,15 @@
-export const NoopLogger = {
+import { Logger } from "./logger";
+
+export type LoggerLike = Pick<
+  Logger,
+  "trace" | "debug" | "info" | "warn" | "error" | "fatal"
+>;
+
+export const NoopLogger: LoggerLike = {
   trace() {},
   debug() {},
   info() {},
   warn() {},
   error() {},
   fatal() {},
-} as const;
+};